+++ inst/app/app_ez/global.R
@@ -1,6 +1,5 @@
-## use shiny::runApp() in console or click right top button
-## '>Run App' in Rstudio to start app,
-## but do not write `shiny::runApp()` in script, type it in console
+## use shiny::runApp() in console or click the top-right button '>Run App'
+## in Rstudio to start app, but do not write `shiny::runApp()` in this script, type it in console
 
 time_start <- Sys.time()
 library(systemPipeShiny)
@@ -8,22 +7,8 @@ library(magrittr) # load pipes
 # load additional libraries that you want to use below
 
 
-## SPS options
+## SPS options to configure the app
 # read "https://systempipe.org/sps/adv_features/config/#app-options" for details
-# title: dashboard and website title - any string
-# title_logo: logo to display when dashboard is collapsed and on browser tab - path/url of an image
-# mode: running mode - "local", "server"
-# warning_toast: some warning messages to check potential risks - TRUE, FALSE
-# login_screen: to show login screen? - TRUE, FALSE
-# login_theme: login screen themes, login_screen need be TRUE - "random", "empty", "vhelix", "hhelix", "biomatrix"
-# use_crayon: Do you want colorful terminal messages? TRUE, FALSE
-# verbose: display some info during processing? - TRUE, FALSE
-# admin_url: admin_page query url - "admin"
-# note_url: User notification broadcast file url - http(s) address
-# tab_welcome -- module_wf: whether to load the corresponding tab or module? - TRUE, FALSE
-# traceback: for expressions wrapped inside `spsComps::shinyCatch`, show full traceback if error? TRUE, FALSE
-# is_demo: useful if deploy the app as a demo - TRUE, FALSE
-# welcome_guide: enable the welcome guide which show you where is guide dropdown menu? - TRUE, FALSE
 spsOption(.list = list(
     title = "systemPipeShiny",
     title_logo = "img/sps_small.png",
@@ -48,7 +33,7 @@ spsOption(.list = list(
     welcome_guide = TRUE
 ))
 
-## An alternative is to comment above and use `spsOption` to overwrite single options, eg:
+## An alternative is to comment above and use `spsOption` to overwrite **single** options, eg:
 # spsOption("mode", "server")
 
 ## use `spsOptions` to check current settings
@@ -77,9 +62,13 @@ options(shiny.maxRequestSize = 30*1e6)
 ####### SPS Main App Function Starts #########
 
 sps_app <- sps(
-    tabs = c("vs_example"),
+    # once you have create more custom tabs, add IDs of these tabs in this vector
+    # to confirm you want to load them.
+    tabs = c(
+        "vs_example"
+    ),
     server_expr = {
-        # add you own server functions below
+        # add you own additional server expressions below, for example:
         msg("Custom expression runs -- Hello World", "GREETING", "green")
     }
 )